Merge Alex Matveevs change:
This change should be done both in U-Boot and Linux PCI controller
drivers. Reset in U-Boot is unnecessary, as the controller was just
powered on. Reset in Linux is harmful, because it leads to inconsistent
state of PCI devices. This effect is observed with video controller: any
attempt to access it after the reset locks up the system.